LPS-142279 Remove ".jar/" case since it is no longer applicable
This if condition was originally introduced in LPS-13880 in PropsUtil because JBoss jar URL did not contain "!/". ".jar/" was used instead to get the directory path. Since we moved this method from PropsUtil to ClassUtil which requires full code coverage. We re-evaluated and found that newer JBoss URL now contains "!/". This means we can remove the ".jar/" condition and fix our code coverage.