LPS-190886 Create test to ensure that all existing caches can be reconfigured.
Right now the code is susceptible of the following situation:

1. FinderCache for a model object is removed and created with the new configuration.
2. EntityCache is removed for the same model object.
3. The CacheManager will notify EntityCacheImpl that the cache was removed and method notifyPortalCacheRemoved will be called.
4. Method _notifyFinderCache will be called and the cache created in the first step will be removed again losing its new configuration.


One important thing must be taken into account:

1.The number of the caches has to be big enough (that's why all of them are taken so that the issue is statistically reproducible in a consistent way each time).

The reason why it cannot be ensured its reproduction otherwise is because net.sf.ehcache.config.Configuration.getCacheConfigurations() is a ConcurrentHashMap, so cacheConfigurations.values() order cannot be assumed, and order is importan to reproduce the issue.

The issue is reproducible because at least one cache is removed in the cache manager and not recreated so it cannot be found in the MBeanServer.