LPS-196123 Move class DLPreviewableProcessor and its subclasses into module
# breaking_change_report

## What portal-impl/src/com/liferay/portlet/documentlibrary/util/DLPreviewableProcessor.java

Class DLPreviewableProcessor is moved from portal-impl into document-library-preview-api.

## Why

We want to make AudioProcessorImpl and VideoProcessorImpl components of OSGI, all subclasses of DLPreviewableProcessor will be in the module, so there is no need for it to be in portal-impl.

## Alternatives

Add document-library-preview-api in build dependency and use the same class.

----

# breaking_change_report

## What portal-impl/src/com/liferay/portlet/documentlibrary/util/AudioProcessorImpl.java

Class AudioProcessorImpl is moved from portal-impl into document-library-preview-audio.

## Why

We want to make AudioProcessorImpl a component of OSGI.

## Alternatives

Reference class DLProcessor with target "(type=" + DLProcessorConstants.AUDIO_PROCESSOR + ")".

----

# breaking_change_report

## What portal-impl/src/com/liferay/portlet/documentlibrary/util/VideoProcessorImpl.java

Class VideoProcessorImpl is moved from portal-impl into document-library-preview-video.

## Why

We want to make VideoProcessorImpl a component of OSGI.

## Alternatives

Reference class DLProcessor with target "(type=" + DLProcessorConstants.VIDEO_PROCESSOR + ")".

----