LPD-22268 Add copyright to generated code
# breaking

## What portal-impl/src/com/liferay/portal/tools/InstanceWrapperBuilder.java

Add new parameter to constructor method so that we can still add
Liferay's copyright to the generated code.

## Why

We want to remove the usage of Jalopy from InstanceWrapperBuilder as we
are trying to remove it from the codebase. Since Jalopy was responsible
for adding the copyright in the generated code, we must now manually do
this.

----