LPD-40015 set the email address directly after verifying and update the user model
Most of the update methods for user models check for
company.isStrangersVerify and will send a verification email rather than updating
the user's email address.  This doesn't make sense when creating the
instance with the API if the admin's email is supplied.

We can avoid changing the update method used if we update the model with the
new email first and still be able verify the other fields.