LPD-42697 Semantic versioning
# breaking

## What portal-kernel/src/com/liferay/portal/kernel/util/PropsKeys.java

Remove PropsKeys.WORK_DIR_OVERRIDE_ENABLED

## Why

This is a developer tool feature to allow quick deployment of JSP/JS/CSS to server. However it is not widely used, its complexity is causing troubles on migrating Jasper codebase for Jakarta support and it affects performance. Developer tool can still do the normal osgi bundle deployment to fulfill the same function, so removing this feature to simplify codebase.

----

## What portal-impl/src/com/liferay/portal/util/PropsValues.java

Remove PropsValues.WORK_DIR_OVERRIDE_ENABLED

## Why

This is a developer tool feature to allow quick deployment of JSP/JS/CSS to server. However it is not widely used, its complexity is causing troubles on migrating Jasper codebase for Jakarta support and it affects performance. Developer tool can still do the normal osgi bundle deployment to fulfill the same function, so removing this feature to simplify codebase.

----