LPD-61152 Add unit test to expose issue
In the "replace()" method of ServiceBag.java, even though
"currentService" and "_serviceWrapper" point to the same service wrapper
object, each one is contained in a different proxy, causing the equality
check to break. This happens for classes like LayoutLocalService and
LayoutSetLocalService that directly set AopInvocationHandler via
LayoutLocalServiceStagingAdvice and LayoutSetLocalServiceStagingAdvice
respectively.