LPD-72034 Remove HTTP protocol configuration from S3Store
With AWS SDK v2, there is no direct equivalent for the 'connectionProtocol'
property. The alternative is when the S3 store is configured to use an endpoint
via the 's3Endpoint' configuration property, the 's3Endpoint' value can be used
to specify the connection protocol. For example, 'https://example.com' specifies
that HTTPS should be used while 'http://example.com' specifies that HTTP should
be used.